+++ _config.yml
@@ -1,5 +1,5 @@
-title: CasaXML
-description: The swiss real-estate publishing standard that makes sense
+title: SwissRETS
+description: A swiss real estate transfer standard
 google_analytics: 
 show_downloads: true
 theme: jekyll-theme-minimal
