+++ scripts/update/lib/exec.js
